refactor(qa): remove gateway child test facade
