+++ extensions/qa-lab/src/cli-paths.ts
@@ -1,5 +1,4 @@
 // Qa Lab plugin module implements cli paths behavior.
-import fs from "node:fs/promises";
 import path from "node:path";
 import { assertNoSymlinkParents, pathScope } from "openclaw/plugin-sdk/security-runtime";
 
@@ -29,25 +28,6 @@ export function resolveRepoRelativeOutputDir(repoRoot: string, outputDir?: strin
   return resolved.path;
 }
 
-async function resolveNearestExistingPath(targetPath: string) {
-  let current = path.resolve(targetPath);
-  while (true) {
-    try {
-      await fs.lstat(current);
-      return current;
-    } catch (error) {
-      if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
-        throw error;
-      }
-    }
-    const parent = path.dirname(current);
-    if (parent === current) {
-      throw new Error(`failed to resolve existing path for ${targetPath}`);
-    }
-    current = parent;
-  }
-}
-
 function assertRepoRelativePath(repoRoot: string, targetPath: string, label: string) {
   const relative = path.relative(repoRoot, targetPath);
   if (relative.startsWith("..") || path.isAbsolute(relative)) {
@@ -70,18 +50,6 @@ async function assertNoSymlinkSegments(repoRoot: string, targetPath: string, lab
     }
     throw error;
   }
-}
-
-export async function assertRepoBoundPath(repoRoot: string, targetPath: string, label: string) {
-  const repoRootResolved = path.resolve(repoRoot);
-  const targetResolved = path.resolve(targetPath);
-  assertRepoRelativePath(repoRootResolved, targetResolved, label);
-  await assertNoSymlinkSegments(repoRootResolved, targetResolved, label);
-  const repoRootReal = await fs.realpath(repoRootResolved);
-  const nearestExistingPath = await resolveNearestExistingPath(targetResolved);
-  const nearestExistingReal = await fs.realpath(nearestExistingPath);
-  assertRepoRelativePath(repoRootReal, nearestExistingReal, label);
-  return targetResolved;
 }
 
 export async function ensureRepoBoundDirectory(
