+++ extensions/qa-lab/src/gateway-child.test.ts
@@ -9,12 +9,41 @@ import { pathToFileURL } from "node:url";
 import { toErrorObject } from "openclaw/plugin-sdk/error-runtime";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import {
-  testing,
+  createQaBundledPluginsDir,
+  resolveQaOwnerPluginIdsForProviderIds,
+  resolveQaRuntimeHostVersion,
+} from "./bundled-plugin-staging.js";
+import { preserveQaGatewayDebugArtifacts } from "./gateway-child-artifacts.js";
+import { resolveQaGatewayChildCommand, runQaGatewayCliCommand } from "./gateway-child-command.js";
+import {
+  buildQaForcedRuntimeEnvPatch,
   buildQaRuntimeEnv,
-  resolveQaControlUiRoot,
-  startQaGatewayChild,
-} from "./gateway-child.js";
+  stageQaCodexMockModelCatalog,
+} from "./gateway-child-env.js";
+import {
+  closeQaGatewayLogStream,
+  createQaGatewayChildLogCollector,
+  formatQaGatewayProcessBoundaryStartupFailure,
+  monitorQaGatewayChildFailure,
+  stopQaGatewayChildProcessTree,
+  throwQaGatewayChildFailure,
+} from "./gateway-child-process.js";
+import {
+  callQaGatewayWithRetry,
+  isRetryableRpcStartupError,
+  resolveQaGatewayStartupRetry,
+  waitForGatewayReady,
+  waitForQaGatewayRestartBoundary,
+} from "./gateway-child-readiness.js";
+import { startQaGatewayChild } from "./gateway-child.js";
+import { readQaLiveProviderConfigOverrides } from "./providers/live-config.js";
+import {
+  assertQaLiveCodexAuthAvailable,
+  stageQaLiveAnthropicSetupToken,
+  stageQaLiveApiKeyProfiles,
+} from "./providers/live-frontier/auth.js";
 import { readQaAuthProfiles } from "./providers/shared/auth-store.js";
+import { stageQaMockAuthProfiles } from "./providers/shared/mock-auth.js";
 import { createTempDirHarness } from "./temp-dir.test-helper.js";
 
 const fetchWithSsrFGuardMock = vi.hoisted(() => vi.fn());
@@ -105,16 +134,6 @@ function requireSsrFetchCall(index = 0): SsrFetchCall {
   return call[0] as SsrFetchCall;
 }
 
-async function expectPathMissing(filePath: string): Promise<void> {
-  try {
-    await lstat(filePath);
-  } catch (error) {
-    expect((error as NodeJS.ErrnoException).code).toBe("ENOENT");
-    return;
-  }
-  throw new Error(`expected ${filePath} to be missing`);
-}
-
 async function writeJsonFixture(filePath: string, value: unknown, space?: number) {
   await mkdir(path.dirname(filePath), { recursive: true });
   await writeFile(filePath, JSON.stringify(value, null, space), "utf8");
@@ -187,7 +206,7 @@ async function readJsonLines(filePath: string): Promise<Array<Record<string, unk
 
 describe("runQaGatewayCliCommand", () => {
   it("runs CLI commands with the Gateway fixture environment", async () => {
-    const output = await testing.runQaGatewayCliCommand({
+    const output = await runQaGatewayCliCommand({
       executablePath: process.execPath,
       argsPrefix: [
         "--eval",
@@ -203,7 +222,7 @@ describe("runQaGatewayCliCommand", () => {
 
   it("reports CLI stderr when a fixture command fails", async () => {
     await expect(
-      testing.runQaGatewayCliCommand({
+      runQaGatewayCliCommand({
         executablePath: process.execPath,
         argsPrefix: ["--eval", 'process.stderr.write("fixture failure"); process.exit(7)'],
         args: [],
@@ -212,25 +231,6 @@ describe("runQaGatewayCliCommand", () => {
       }),
     ).rejects.toThrow("OpenClaw CLI exited 7: fixture failure");
   });
-
-  it.each(["stdout", "stderr"] as const)(
-    "rejects and stops the CLI child when its %s pipe fails",
-    async (streamName) => {
-      const child = spawn(process.execPath, ["--eval", "setInterval(() => {}, 1000)"], {
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-      const close = once(child, "close");
-      const result = testing.readQaGatewayCliCommand(child);
-      const message = `synthetic ${streamName} read failure`;
-
-      child[streamName]?.destroy(new Error(message));
-
-      await expect(result).rejects.toThrow(
-        `qa gateway cli ${streamName} stream failed: ${message}`,
-      );
-      await close;
-    },
-  );
 });
 
 describe("monitorQaGatewayChildFailure", () => {
@@ -240,8 +240,8 @@ describe("monitorQaGatewayChildFailure", () => {
       stdio: ["ignore", "pipe", "pipe"],
     });
     const close = once(child, "close");
-    const output = testing.createQaGatewayChildLogCollector();
-    const getFailure = testing.monitorQaGatewayChildFailure(child, output);
+    const output = createQaGatewayChildLogCollector();
+    const getFailure = monitorQaGatewayChildFailure(child, output);
     const error = new Error("synthetic gateway stdout read failure");
 
     child.stdout?.destroy(error);
@@ -253,7 +253,7 @@ describe("monitorQaGatewayChildFailure", () => {
       "gateway child stdout stream failed: synthetic gateway stdout read failure",
     );
     expect(output.text()).not.toContain("later stderr read failure");
-    expect(() => testing.throwQaGatewayChildFailure(getFailure, () => output.text())).toThrow(
+    expect(() => throwQaGatewayChildFailure(getFailure, () => output.text())).toThrow(
       "gateway child stdout stream failed: synthetic gateway stdout read failure",
     );
   });
@@ -263,7 +263,7 @@ describe("formatQaGatewayProcessBoundaryStartupFailure", () => {
   it("includes only a bounded, redacted launcher log tail", () => {
     const prefix = "x".repeat(9_000);
     const longSecret = "s".repeat(9_000);
-    const message = testing.formatQaGatewayProcessBoundaryStartupFailure(
+    const message = formatQaGatewayProcessBoundaryStartupFailure(
       new Error("launcher exited before identity"),
       `${prefix}\nAuthorization: Bearer ${longSecret}\nlauncher stage=mount-proc`,
     );
@@ -277,7 +277,7 @@ describe("formatQaGatewayProcessBoundaryStartupFailure", () => {
   });
 
   it("preserves complete Unicode code points at the retained log-tail boundary", () => {
-    const message = testing.formatQaGatewayProcessBoundaryStartupFailure(
+    const message = formatQaGatewayProcessBoundaryStartupFailure(
       new Error("launcher exited before identity"),
       `P😀${"z".repeat(8_191)}`,
     );
@@ -304,7 +304,7 @@ describe("waitForGatewayReady", () => {
       });
 
       try {
-        const readiness = testing.waitForGatewayReady({
+        const readiness = waitForGatewayReady({
           baseUrl,
           logs: () => `${phase} logs`,
           child: { exitCode: null, signalCode: null },
@@ -316,6 +316,11 @@ describe("waitForGatewayReady", () => {
         expect(fetchWithSsrFGuardMock.mock.calls.map(([request]) => request.url)).toEqual([
           `${baseUrl}/readyz`,
         ]);
+        const healthRequest = requireSsrFetchCall();
+        expect(healthRequest.init?.method).toBe("HEAD");
+        expect(healthRequest.init?.headers).toEqual({ connection: "close" });
+        expect(healthRequest.policy).toEqual({ allowPrivateNetwork: true });
+        expect(healthRequest.auditContext).toBe("qa-lab-gateway-child-health");
         expect(release).toHaveBeenCalledTimes(1);
 
         ready = true;
@@ -349,7 +354,7 @@ describe("waitForGatewayReady", () => {
     const startedAt = Date.now();
 
     await expect(
-      testing.waitForGatewayReady({
+      waitForGatewayReady({
         baseUrl: "http://127.0.0.1:43124",
         logs: () => "near-expiry logs",
         child: { exitCode: null, signalCode: null },
@@ -363,16 +368,9 @@ describe("waitForGatewayReady", () => {
 });
 
 describe("Gateway child fixture helpers", () => {
-  it("creates an empty transport config seam", () => {
-    expect(testing.createQaGatewayEmptyTransport()).toEqual({
-      requiredPluginIds: [],
-      createGatewayConfig: expect.any(Function),
-    });
-  });
-
   it("stages native Codex model metadata before starting the private mock runtime", async () => {
     const tempRoot = await tempDirs.makeTempDir("qa-codex-model-catalog-");
-    const modelCatalogPath = await testing.stageQaCodexMockModelCatalog({
+    const modelCatalogPath = await stageQaCodexMockModelCatalog({
       tempRoot,
       forcedRuntime: "codex",
       providerMode: "mock-openai",
@@ -388,16 +386,19 @@ describe("Gateway child fixture helpers", () => {
       expect.objectContaining({
         slug: "gpt-5.6-luna",
         apply_patch_tool_type: "freeform",
+        supports_reasoning_summary_parameter: true,
         tool_mode: "direct",
       }),
       expect.objectContaining({
         slug: "gpt-5.6-luna-alt",
         apply_patch_tool_type: "freeform",
+        supports_reasoning_summary_parameter: true,
         tool_mode: "direct",
       }),
     ]);
+    expect(catalog.models[0]).not.toHaveProperty("supports_reasoning_summaries");
     expect(
-      testing.buildQaForcedRuntimeEnvPatch({
+      buildQaForcedRuntimeEnvPatch({
         forcedRuntime: "codex",
         providerMode: "mock-openai",
         providerBaseUrl: "http://127.0.0.1:44080/v1",
@@ -413,14 +414,14 @@ describe("Gateway child fixture helpers", () => {
   it("does not stage a Codex catalog for other runtimes or live providers", async () => {
     const tempRoot = await tempDirs.makeTempDir("qa-codex-model-catalog-unused-");
     await expect(
-      testing.stageQaCodexMockModelCatalog({
+      stageQaCodexMockModelCatalog({
         tempRoot,
         forcedRuntime: "openclaw",
         providerMode: "mock-openai",
       }),
     ).resolves.toBeUndefined();
     await expect(
-      testing.stageQaCodexMockModelCatalog({
+      stageQaCodexMockModelCatalog({
         tempRoot,
         forcedRuntime: "codex",
         providerMode: "live-frontier",
@@ -431,46 +432,13 @@ describe("Gateway child fixture helpers", () => {
     ).rejects.toThrow();
   });
 
-  it("confines live Codex QA without replacing its native provider configuration", () => {
-    expect(
-      testing.buildQaForcedRuntimeEnvPatch({
-        forcedRuntime: "codex",
-        providerMode: "live-frontier",
-      }),
-    ).toEqual({
-      OPENCLAW_BUILD_PRIVATE_QA: "1",
-      OPENCLAW_QA_FORCE_RUNTIME: "codex",
-      OPENCLAW_CODEX_APP_SERVER_ARGS:
-        "app-server -c sandbox_workspace_write.exclude_tmpdir_env_var=true " +
-        "-c sandbox_workspace_write.exclude_slash_tmp=true --listen stdio://",
-    });
-  });
-
-  it("preserves preconfigured live Codex arguments while enforcing QA containment", () => {
-    expect(
-      testing.buildQaForcedRuntimeEnvPatch({
-        forcedRuntime: "codex",
-        providerMode: "live-frontier",
-        nativeAppServerArgs:
-          'app-server -c openai_base_url="https://live.example/v1" --listen stdio://',
-      }),
-    ).toEqual({
-      OPENCLAW_BUILD_PRIVATE_QA: "1",
-      OPENCLAW_QA_FORCE_RUNTIME: "codex",
-      OPENCLAW_CODEX_APP_SERVER_ARGS:
-        'app-server -c openai_base_url="https://live.example/v1" --listen stdio:// ' +
-        "-c sandbox_workspace_write.exclude_tmpdir_env_var=true " +
-        "-c sandbox_workspace_write.exclude_slash_tmp=true",
-    });
-  });
-
   it("resolves the repo runner before a built Gateway CLI fallback", async () => {
     const repoRoot = await tempDirs.makeTempDir("qa-gateway-command-");
     await mkdir(path.join(repoRoot, "scripts"), { recursive: true });
     const runnerPath = path.join(repoRoot, "scripts", "run-node.mjs");
     await writeFile(runnerPath, "export {};\n", "utf8");
 
-    expect(testing.resolveQaGatewayChildCommand(repoRoot)).toEqual({
+    expect(resolveQaGatewayChildCommand(repoRoot)).toEqual({
       executablePath: process.execPath,
       argsPrefix: [runnerPath],
       cwd: repoRoot,
@@ -480,7 +448,7 @@ describe("Gateway child fixture helpers", () => {
     await mkdir(path.join(repoRoot, "dist"), { recursive: true });
     await writeFile(path.join(repoRoot, "dist", "index.js"), "export {};\n", "utf8");
     await rm(path.join(repoRoot, "scripts"), { recursive: true });
-    expect(testing.resolveQaGatewayChildCommand(repoRoot)).toEqual({
+    expect(resolveQaGatewayChildCommand(repoRoot)).toEqual({
       executablePath: process.execPath,
       argsPrefix: [path.join(repoRoot, "dist", "index.js")],
       cwd: repoRoot,
@@ -677,11 +645,6 @@ describe("buildQaRuntimeEnv", () => {
     expect(env.OPENAI_API_KEY).toBe("openai-live");
     expect(env.ANTHROPIC_API_KEY).toBe("anthropic-live");
     expect(env.GEMINI_API_KEY).toBe("gemini-live");
-  });
-
-  it("defaults gateway-child provider mode to mock-openai when omitted", () => {
-    expect(testing.resolveQaGatewayChildProviderMode(undefined)).toBe("mock-openai");
-    expect(testing.resolveQaGatewayChildProviderMode("live-frontier")).toBe("live-frontier");
   });
 
   it("keeps explicit provider env vars over live aliases", () => {
@@ -959,15 +922,6 @@ describe("buildQaRuntimeEnv", () => {
     },
   );
 
-  it("treats restart socket closures as retryable gateway call errors", () => {
-    expect(testing.isRetryableGatewayCallError("gateway closed (1006 abnormal closure)")).toBe(
-      true,
-    );
-    expect(testing.isRetryableGatewayCallError("gateway closed (1012 service restart)")).toBe(true);
-    expect(testing.isRetryableGatewayCallError("service restart in progress")).toBe(true);
-    expect(testing.isRetryableGatewayCallError("permission denied")).toBe(false);
-  });
-
   it("preserves relative gateway retry timeouts without an absolute deadline", async () => {
     const request = vi
       .fn()
@@ -976,7 +930,7 @@ describe("buildQaRuntimeEnv", () => {
     const waitForReady = vi.fn(async () => {});
 
     await expect(
-      testing.callQaGatewayWithRetry({
+      callQaGatewayWithRetry({
         logs: () => "qa logs",
         request,
         throwChildFailure: vi.fn(),
@@ -1002,7 +956,7 @@ describe("buildQaRuntimeEnv", () => {
     });
 
     await expect(
-      testing.callQaGatewayWithRetry({
+      callQaGatewayWithRetry({
         deadlineMs: 10_000,
         logs: () => "qa logs",
         request,
@@ -1021,7 +975,7 @@ describe("buildQaRuntimeEnv", () => {
   it("waits for a fresh in-process restart boundary after the current log offset", async () => {
     let logs = "old restart mode: in-process restart\n";
     const mark = logs.length;
-    const wait = testing.waitForQaGatewayRestartBoundary({
+    const wait = waitForQaGatewayRestartBoundary({
       readLogsSince: (since) => logs.slice(since),
       mark,
       pollMs: 1,
@@ -1034,11 +988,11 @@ describe("buildQaRuntimeEnv", () => {
   });
 
   it("keeps restart offsets stable after stderr output", async () => {
-    const output = testing.createQaGatewayChildLogCollector();
+    const output = createQaGatewayChildLogCollector();
     output.push("stdout", Buffer.from("gateway ready\n"));
     output.push("stderr", Buffer.from("stderr warning\n"));
     const mark = output.mark();
-    const wait = testing.waitForQaGatewayRestartBoundary({
+    const wait = waitForQaGatewayRestartBoundary({
       readLogsSince: (since) => output.readSince(since),
       mark,
       pollMs: 1,
@@ -1054,7 +1008,7 @@ describe("buildQaRuntimeEnv", () => {
   });
 
   it("bounds diagnostics while monotonic marks retain fresh output semantics", () => {
-    const output = testing.createQaGatewayChildLogCollector();
+    const output = createQaGatewayChildLogCollector();
     output.push("stdout", Buffer.from(`old😀${"x".repeat(70_000)}`));
     const mark = output.mark();
     output.push("stdout", Buffer.from("fresh restart mode: in-process restart\n"));
@@ -1068,7 +1022,7 @@ describe("buildQaRuntimeEnv", () => {
   });
 
   it("decodes interleaved stdout and stderr independently", () => {
-    const output = testing.createQaGatewayChildLogCollector();
+    const output = createQaGatewayChildLogCollector();
     const stdout = Buffer.from("before 😀 after\n");
 
     output.push("stdout", stdout.subarray(0, 9));
@@ -1081,7 +1035,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("times out when a SIGUSR1 restart never reaches the boundary", async () => {
     await expect(
-      testing.waitForQaGatewayRestartBoundary({
+      waitForQaGatewayRestartBoundary({
         readLogsSince: () => "signal SIGUSR1 received\n",
         mark: 0,
         pollMs: 1,
@@ -1092,7 +1046,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("keeps oversized restart-boundary poll intervals within the timeout", async () => {
     await expect(
-      testing.waitForQaGatewayRestartBoundary({
+      waitForQaGatewayRestartBoundary({
         readLogsSince: () => "signal SIGUSR1 received\n",
         mark: 0,
         pollMs: Number.MAX_SAFE_INTEGER,
@@ -1105,7 +1059,7 @@ describe("buildQaRuntimeEnv", () => {
     const stateDir = await tempDirs.makeTempDir("qa-setup-token-state-");
     const token = `sk-ant-oat01-${"c".repeat(80)}`;
 
-    const cfg = await testing.stageQaLiveAnthropicSetupToken({
+    const cfg = await stageQaLiveAnthropicSetupToken({
       cfg: {},
       stateDir,
       env: {
@@ -1128,7 +1082,7 @@ describe("buildQaRuntimeEnv", () => {
   it("stages live env API-key profiles for isolated QA workers", async () => {
     const stateDir = await tempDirs.makeTempDir("qa-live-api-key-state-");
 
-    const cfg = await testing.stageQaLiveApiKeyProfiles({
+    const cfg = await stageQaLiveApiKeyProfiles({
       cfg: {},
       stateDir,
       providerIds: ["openai"],
@@ -1157,45 +1111,10 @@ describe("buildQaRuntimeEnv", () => {
     }
   });
 
-  it("stages the OpenAI API-key fallback for live OpenAI QA workers", async () => {
-    const stateDir = await tempDirs.makeTempDir("qa-live-codex-api-key-state-");
-
-    const cfg = await testing.stageQaLiveApiKeyProfiles({
-      cfg: {},
-      stateDir,
-      providerIds: ["openai"],
-      env: {
-        OPENCLAW_LIVE_OPENAI_KEY: "qa-live-codex-fallback-key",
-      },
-    });
-
-    for (const [profileId, provider] of [
-      ["qa-live-openai-env", "openai"],
-      ["qa-live-openai-env", "openai"],
-    ] as const) {
-      const configProfile = requireAuthProfile(cfg.auth?.profiles, profileId);
-      expect(configProfile.provider).toBe(provider);
-      expect(configProfile.mode).toBe("api_key");
-    }
-
-    for (const agentId of ["main", "qa"]) {
-      const storeProfiles = readAuthProfileStore(stateDir, agentId).profiles;
-      for (const [profileId, provider] of [
-        ["qa-live-openai-env", "openai"],
-        ["qa-live-openai-env", "openai"],
-      ] as const) {
-        const storeProfile = requireAuthProfile(storeProfiles, profileId);
-        expect(storeProfile.type).toBe("api_key");
-        expect(storeProfile.provider).toBe(provider);
-        expect(storeProfile.key).toBe("qa-live-codex-fallback-key");
-      }
-    }
-  });
-
   it("stages direct live OpenAI API-key aliases for isolated QA workers", async () => {
     const stateDir = await tempDirs.makeTempDir("qa-live-codex-direct-key-state-");
 
-    const cfg = await testing.stageQaLiveApiKeyProfiles({
+    const cfg = await stageQaLiveApiKeyProfiles({
       cfg: {},
       stateDir,
       providerIds: ["openai"],
@@ -1213,7 +1132,7 @@ describe("buildQaRuntimeEnv", () => {
     expect(storeProfile.key).toBe("qa-live-direct-codex-key");
 
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg,
         providerIds: ["openai"],
         env: {
@@ -1226,20 +1145,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("fails fast when live OpenAI runs have no portable QA auth", () => {
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
-        cfg: {},
-        providerIds: ["openai"],
-        env: {
-          CODEX_HOME: path.join(os.tmpdir(), "missing-openclaw-codex-home"),
-        },
-        readCodexCredentials: () => null,
-      }),
-    ).toThrow("QA live-frontier cannot run Codex-backed OpenAI models");
-  });
-
-  it("fails fast when default OpenAI model refs route through Codex without portable QA auth", () => {
-    expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg: {},
         providerIds: ["openai"],
         env: {
@@ -1252,7 +1158,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("does not require Codex auth for custom OpenAI-compatible provider configs", () => {
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg: {
           models: {
             providers: {
@@ -1274,7 +1180,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("fails fast when forced Codex runtime uses OpenAI model refs without portable QA auth", () => {
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg: {},
         providerIds: ["openai"],
         env: {
@@ -1288,7 +1194,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("accepts OpenAI API-key fallback auth for forced Codex runtime QA runs", () => {
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg: {},
         providerIds: ["openai"],
         env: {
@@ -1302,7 +1208,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("stages configured OpenAI API keys for live QA runs", async () => {
     const stateDir = await tempDirs.makeTempDir("qa-live-codex-config-key-state-");
-    const cfg = await testing.stageQaLiveApiKeyProfiles({
+    const cfg = await stageQaLiveApiKeyProfiles({
       cfg: {
         models: {
           providers: {
@@ -1333,7 +1239,7 @@ describe("buildQaRuntimeEnv", () => {
     }
 
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg,
         providerIds: ["openai"],
         env: {},
@@ -1347,7 +1253,7 @@ describe("buildQaRuntimeEnv", () => {
     const env = {
       OPENCLAW_LIVE_CODEX_API_KEY: "qa-configured-env-ref-not-a-real-key",
     };
-    const cfg = await testing.stageQaLiveApiKeyProfiles({
+    const cfg = await stageQaLiveApiKeyProfiles({
       cfg: {
         models: {
           providers: {
@@ -1377,7 +1283,7 @@ describe("buildQaRuntimeEnv", () => {
     expect(storeProfile.key).toBe("qa-configured-env-ref-not-a-real-key");
 
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg,
         providerIds: ["openai"],
         env,
@@ -1388,7 +1294,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("stages configured OpenAI env markers for live QA runs", async () => {
     const stateDir = await tempDirs.makeTempDir("qa-live-codex-config-marker-state-");
-    const cfg = await testing.stageQaLiveApiKeyProfiles({
+    const cfg = await stageQaLiveApiKeyProfiles({
       cfg: {
         models: {
           providers: {
@@ -1416,7 +1322,7 @@ describe("buildQaRuntimeEnv", () => {
     expect(storeProfile.key).toBe("qa-configured-marker-not-a-real-key");
 
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg,
         providerIds: ["openai"],
         env: {},
@@ -1435,7 +1341,7 @@ describe("buildQaRuntimeEnv", () => {
     }));
 
     expect(() =>
-      testing.assertQaLiveCodexAuthAvailable({
+      assertQaLiveCodexAuthAvailable({
         cfg: {},
         providerIds: ["openai"],
         env: {
@@ -1454,7 +1360,7 @@ describe("buildQaRuntimeEnv", () => {
   it("stages placeholder mock auth profiles per agent dir so mock-openai runs can resolve credentials", async () => {
     const stateDir = await tempDirs.makeTempDir("qa-mock-auth-");
 
-    const cfg = await testing.stageQaMockAuthProfiles({
+    const cfg = await stageQaMockAuthProfiles({
       cfg: {},
       stateDir,
     });
@@ -1591,7 +1497,7 @@ describe("buildQaRuntimeEnv", () => {
   it("stages mock profiles only for the requested agents and providers when callers override the defaults", async () => {
     const stateDir = await tempDirs.makeTempDir("qa-mock-auth-override-");
 
-    const cfg = await testing.stageQaMockAuthProfiles({
+    const cfg = await stageQaMockAuthProfiles({
       cfg: {},
       stateDir,
       agentIds: ["qa"],
@@ -1614,27 +1520,6 @@ describe("buildQaRuntimeEnv", () => {
     await expect(
       lstat(path.join(stateDir, "agents", "main", "agent", "openclaw-agent.sqlite")),
     ).rejects.toThrow(/ENOENT/);
-  });
-
-  it("allows loopback gateway health probes through the SSRF guard", async () => {
-    const release = vi.fn(async () => {});
-    fetchWithSsrFGuardMock.mockResolvedValue({
-      response: { ok: true },
-      release,
-    });
-
-    await expect(
-      testing.fetchLocalGatewayHealth({
-        baseUrl: "http://127.0.0.1:18789",
-        healthPath: "/readyz",
-      }),
-    ).resolves.toBe(true);
-
-    const request = requireSsrFetchCall();
-    expect(request.url).toBe("http://127.0.0.1:18789/readyz");
-    expect(request.policy).toEqual({ allowPrivateNetwork: true });
-    expect(request.auditContext).toBe("qa-lab-gateway-child-health");
-    expect(release).toHaveBeenCalledTimes(1);
   });
 
   it("force-stops gateway children that ignore the graceful signal", async () => {
@@ -1661,8 +1546,8 @@ describe("buildQaRuntimeEnv", () => {
       return true;
     });
 
-    await testing.stopQaGatewayChildProcessTree(
-      child as unknown as Parameters<typeof testing.stopQaGatewayChildProcessTree>[0],
+    await stopQaGatewayChildProcessTree(
+      child as unknown as Parameters<typeof stopQaGatewayChildProcessTree>[0],
       {
         gracefulTimeoutMs: 1,
         forceTimeoutMs: 10,
@@ -1679,22 +1564,6 @@ describe("buildQaRuntimeEnv", () => {
     expect([child.exitCode, child.signalCode]).not.toEqual([null, null]);
   });
 
-  it("lets the gateway finish its bounded shutdown before process-tree escalation", () => {
-    expect(testing.resolveQaGatewayChildStopTimeouts()).toEqual({
-      gracefulTimeoutMs: 30_000,
-      forceTimeoutMs: 10_000,
-    });
-    expect(
-      testing.resolveQaGatewayChildStopTimeouts({
-        gracefulTimeoutMs: 1,
-        forceTimeoutMs: 2,
-      }),
-    ).toEqual({
-      gracefulTimeoutMs: 1,
-      forceTimeoutMs: 2,
-    });
-  });
-
   it("force-closes a gateway log stream whose final flush never settles", async () => {
     const stream = new Writable({
       write(_chunk, _encoding, callback) {
@@ -1706,7 +1575,7 @@ describe("buildQaRuntimeEnv", () => {
     });
     const stderr = vi.spyOn(process.stderr, "write").mockImplementation(() => true);
 
-    await testing.closeWriteStream(stream as never, "stdout", 1);
+    await closeQaGatewayLogStream(stream as never, "stdout", 1);
 
     expect(stream.destroyed).toBe(true);
     expect(stderr).toHaveBeenCalledWith(
@@ -1726,7 +1595,7 @@ describe("buildQaRuntimeEnv", () => {
       vi.spyOn(process, "kill").mockImplementation(() => true);
 
       await expect(
-        testing.stopQaGatewayChildProcessTree(child as never, {
+        stopQaGatewayChildProcessTree(child as never, {
           gracefulTimeoutMs: 1,
           forceTimeoutMs: 1,
           inspectLinuxProcessGroup: () => null,
@@ -1751,7 +1620,7 @@ describe("buildQaRuntimeEnv", () => {
     vi.spyOn(process, "kill").mockImplementation(() => true);
     try {
       await expect(
-        testing.stopQaGatewayChildProcessTree(child as never, {
+        stopQaGatewayChildProcessTree(child as never, {
           gracefulTimeoutMs: 1,
           forceTimeoutMs: 1,
           inspectLinuxProcessGroup: () => ({
@@ -1766,146 +1635,6 @@ describe("buildQaRuntimeEnv", () => {
     } finally {
       if (platformDescriptor) {
         Object.defineProperty(process, "platform", platformDescriptor);
-      }
-    }
-  });
-
-  it("classifies Linux zombie-only process groups as stopped", () => {
-    const inspection = testing.inspectLinuxProcessGroupStats(123, [
-      "123 (gateway child) Z 1 123 123 0 -1 0",
-      "124 (helper (worker)) X 1 123 123 0 -1 0",
-      "125 (unrelated) S 1 999 999 0 -1 0",
-    ]);
-
-    expect(inspection).toEqual({
-      alive: false,
-      diagnostics:
-        'pgid=123 members=[pid=123 state=Z command="gateway child", pid=124 state=X command="helper (worker)"]',
-    });
-  });
-
-  it("classifies Linux process groups with a runnable descendant as alive", () => {
-    const inspection = testing.inspectLinuxProcessGroupStats(123, [
-      "123 (gateway child) Z 1 123 123 0 -1 0",
-      "126 (live helper) D 1 123 123 0 -1 0",
-    ]);
-
-    expect(inspection).toEqual({
-      alive: true,
-      diagnostics:
-        'pgid=123 members=[pid=123 state=Z command="gateway child", pid=126 state=D command="live helper"]',
-    });
-  });
-
-  it("classifies an empty Linux process-group snapshot as unknown", () => {
-    expect(
-      testing.inspectLinuxProcessGroupStats(123, ["125 (unrelated) S 1 999 999 0 -1 0"]),
-    ).toEqual({
-      alive: null,
-      diagnostics: "pgid=123 members=[]",
-    });
-  });
-
-  it("bounds Linux process-group diagnostics", () => {
-    const stats = Array.from(
-      { length: 300 },
-      (_, index) => `${index + 1} (${`worker-${index}`.padEnd(32, "x")}) S 1 123 123 0 -1 0`,
-    );
-
-    const inspection = testing.inspectLinuxProcessGroupStats(123, stats);
-
-    expect(inspection.alive).toBe(true);
-    expect(inspection.diagnostics.length).toBeLessThanOrEqual(2_048);
-    expect(inspection.diagnostics).toMatch(/\.\.\.$/u);
-  });
-
-  it("trusts Linux runnable-member inspection before child exit metadata settles", () => {
-    const platformDescriptor = Object.getOwnPropertyDescriptor(process, "platform");
-    Object.defineProperty(process, "platform", { value: "linux", configurable: true });
-    const processKill = vi.spyOn(process, "kill").mockImplementation(() => true);
-    const child = {
-      pid: 12345,
-      exitCode: null,
-      signalCode: null,
-    };
-    try {
-      expect(
-        testing.isQaGatewayChildProcessTreeAlive(child as never, () => ({
-          alive: false,
-          diagnostics: 'pgid=12345 members=[pid=12345 state=Z command="gateway"]',
-        })),
-      ).toBe(false);
-      expect(
-        testing.isQaGatewayChildProcessTreeAlive(child as never, () => ({
-          alive: true,
-          diagnostics: 'pgid=12345 members=[pid=12346 state=S command="worker"]',
-        })),
-      ).toBe(true);
-      expect(
-        testing.isQaGatewayChildProcessTreeAlive(child as never, () => ({
-          alive: null,
-          diagnostics: "pgid=12345 members=[]",
-        })),
-      ).toBe(true);
-      expect(testing.isQaGatewayChildProcessTreeAlive(child as never, () => null)).toBe(true);
-      expect(processKill).toHaveBeenCalledWith(-12345, 0);
-    } finally {
-      if (platformDescriptor) {
-        Object.defineProperty(process, "platform", platformDescriptor);
-      }
-    }
-  });
-
-  it("force-kills Windows gateway process trees when graceful taskkill fails", () => {
-    const platformDescriptor = Object.getOwnPropertyDescriptor(process, "platform");
-    const originalSystemRoot = process.env.SystemRoot;
-    const originalWindir = process.env.WINDIR;
-    Object.defineProperty(process, "platform", { value: "win32", configurable: true });
-    process.env.SystemRoot = "C:\\Windows";
-    delete process.env.WINDIR;
-    try {
-      const child = Object.assign(new EventEmitter(), {
-        pid: 12345,
-        exitCode: null as number | null,
-        signalCode: null as string | null,
-        kill: vi.fn(),
-      });
-      const runTaskkill = vi
-        .fn()
-        .mockReturnValueOnce({ status: 1 })
-        .mockReturnValueOnce({ status: 0 });
-
-      testing.signalQaGatewayChildProcessTree(
-        child as unknown as Parameters<typeof testing.signalQaGatewayChildProcessTree>[0],
-        "SIGTERM",
-        runTaskkill,
-      );
-
-      const taskkillPath = path.win32.join("C:\\Windows", "System32", "taskkill.exe");
-      expect(runTaskkill).toHaveBeenNthCalledWith(1, taskkillPath, ["/PID", "12345", "/T"], {
-        stdio: "ignore",
-        windowsHide: true,
-        timeout: 5_000,
-      });
-      expect(runTaskkill).toHaveBeenNthCalledWith(2, taskkillPath, ["/PID", "12345", "/T", "/F"], {
-        stdio: "ignore",
-        windowsHide: true,
-        timeout: 5_000,
-      });
-      expect(child.kill).not.toHaveBeenCalled();
-    } finally {
-      if (platformDescriptor) {
-        Object.defineProperty(process, "platform", platformDescriptor);
-      }
-      if (originalSystemRoot === undefined) {
-        delete process.env.SystemRoot;
-      } else {
-        process.env.SystemRoot = originalSystemRoot;
-      }
-      if (originalWindir === undefined) {
-        delete process.env.WINDIR;
-      } else {
-        process.env.WINDIR = originalWindir;
       }
     }
   });
@@ -1933,8 +1662,8 @@ describe("buildQaRuntimeEnv", () => {
       return true;
     });
 
-    await testing.stopQaGatewayChildProcessTree(
-      child as unknown as Parameters<typeof testing.stopQaGatewayChildProcessTree>[0],
+    await stopQaGatewayChildProcessTree(
+      child as unknown as Parameters<typeof stopQaGatewayChildProcessTree>[0],
       {
         gracefulTimeoutMs: 1,
         forceTimeoutMs: 50,
@@ -1955,22 +1684,24 @@ describe("buildQaRuntimeEnv", () => {
     }
   });
 
-  it("classifies bind collisions separately from migration convergence restarts", () => {
+  it.each([
+    ["another gateway instance is already listening on ws://127.0.0.1:43124", "bind-collision"],
+    [
+      "failed to bind gateway socket on ws://127.0.0.1:43124: Error: listen EADDRINUSE",
+      "bind-collision",
+    ],
+    [
+      "OpenClaw plugin migration inputs changed during startup convergence; refusing to report the gateway ready. Restart OpenClaw so state migrations run against the final config and plugin inventory.",
+      "migration-convergence-restart",
+    ],
+  ] as const)("classifies %s", (details, expectedKind) => {
     expect(
-      testing.classifyQaGatewayStartupRetry(
-        "another gateway instance is already listening on ws://127.0.0.1:43124",
-      ),
-    ).toBe("bind-collision");
-    expect(
-      testing.classifyQaGatewayStartupRetry(
-        "failed to bind gateway socket on ws://127.0.0.1:43124: Error: listen EADDRINUSE",
-      ),
-    ).toBe("bind-collision");
-    expect(
-      testing.classifyQaGatewayStartupRetry(
-        "OpenClaw plugin migration inputs changed during startup convergence; refusing to report the gateway ready. Restart OpenClaw so state migrations run against the final config and plugin inventory.",
-      ),
-    ).toBe("migration-convergence-restart");
+      resolveQaGatewayStartupRetry({
+        attempt: 1,
+        details,
+        migrationConvergenceRestartUsed: false,
+      })?.kind,
+    ).toBe(expectedKind);
   });
 
   it.each([
@@ -1979,11 +1710,17 @@ describe("buildQaRuntimeEnv", () => {
     "Restart OpenClaw so state migrations can continue.",
     "gateway failed to become healthy",
   ])("does not retry unrelated startup failure: %s", (details) => {
-    expect(testing.classifyQaGatewayStartupRetry(details)).toBeNull();
+    expect(
+      resolveQaGatewayStartupRetry({
+        attempt: 1,
+        details,
+        migrationConvergenceRestartUsed: false,
+      }),
+    ).toBeNull();
   });
 
   it("restarts migration convergence once with the same launch state", () => {
-    const first = testing.resolveQaGatewayStartupRetry({
+    const first = resolveQaGatewayStartupRetry({
       attempt: 1,
       details:
         "OpenClaw plugin migration inputs changed during startup convergence; refusing readiness.",
@@ -1996,7 +1733,7 @@ describe("buildQaRuntimeEnv", () => {
       migrationConvergenceRestartUsed: true,
     });
     expect(
-      testing.resolveQaGatewayStartupRetry({
+      resolveQaGatewayStartupRetry({
         attempt: 2,
         details:
           "OpenClaw plugin migration inputs changed during startup convergence; refusing readiness.",
@@ -2007,7 +1744,7 @@ describe("buildQaRuntimeEnv", () => {
 
   it("rotates launch state only for a bind collision", () => {
     expect(
-      testing.resolveQaGatewayStartupRetry({
+      resolveQaGatewayStartupRetry({
         attempt: 1,
         details: "listen EADDRINUSE: address already in use",
         migrationConvergenceRestartUsed: false,
@@ -2021,14 +1758,14 @@ describe("buildQaRuntimeEnv", () => {
 
   it("fails immediately for generic exits and after the startup attempt budget", () => {
     expect(
-      testing.resolveQaGatewayStartupRetry({
+      resolveQaGatewayStartupRetry({
         attempt: 1,
         details: "gateway exited with code 1",
         migrationConvergenceRestartUsed: false,
       }),
     ).toBeNull();
     expect(
-      testing.resolveQaGatewayStartupRetry({
+      resolveQaGatewayStartupRetry({
         attempt: 5,
         details: "listen EADDRINUSE",
         migrationConvergenceRestartUsed: false,
@@ -2038,35 +1775,11 @@ describe("buildQaRuntimeEnv", () => {
 
   it("treats startup token mismatches as retryable rpc startup errors", () => {
     expect(
-      testing.isRetryableRpcStartupError(
+      isRetryableRpcStartupError(
         "unauthorized: gateway token mismatch (set gateway.remote.token to match gateway.auth.token)",
       ),
     ).toBe(true);
-    expect(testing.isRetryableRpcStartupError("permission denied")).toBe(false);
-  });
-
-  it("probes gateway health with a one-shot HEAD request through the SSRF guard", async () => {
-    const release = vi.fn(async () => {});
-    fetchWithSsrFGuardMock.mockResolvedValue({
-      response: { ok: true },
-      release,
-    });
-
-    await expect(
-      testing.fetchLocalGatewayHealth({
-        baseUrl: "http://127.0.0.1:43124",
-        healthPath: "/readyz",
-      }),
-    ).resolves.toBe(true);
-
-    const request = requireSsrFetchCall();
-    expect(request.url).toBe("http://127.0.0.1:43124/readyz");
-    expect(request.init?.method).toBe("HEAD");
-    expect(request.init?.headers).toEqual({ connection: "close" });
-    expect(request.init?.signal).toBeInstanceOf(AbortSignal);
-    expect(request.policy).toEqual({ allowPrivateNetwork: true });
-    expect(request.auditContext).toBe("qa-lab-gateway-child-health");
-    expect(release).toHaveBeenCalledTimes(1);
+    expect(isRetryableRpcStartupError("permission denied")).toBe(false);
   });
 
   it("preserves only sanitized gateway debug artifacts", async () => {
@@ -2110,7 +1823,7 @@ describe("buildQaRuntimeEnv", () => {
     await mkdir(path.join(tempRoot, "state"), { recursive: true });
     await writeFile(path.join(tempRoot, "state", "secret.txt"), "do-not-copy", "utf8");
 
-    await testing.preserveQaGatewayDebugArtifacts({
+    await preserveQaGatewayDebugArtifacts({
       preserveToDir: artifactDir,
       stdoutLogPath,
       stderrLogPath,
@@ -2156,143 +1869,9 @@ describe("buildQaRuntimeEnv", () => {
       tempRoot,
     );
   });
-
-  it("rejects preserved gateway artifacts outside the repo root", async () => {
-    await expect(
-      testing.assertQaArtifactDirWithinRepo("/tmp/openclaw-repo", "/tmp/outside"),
-    ).rejects.toThrow("QA gateway artifact directory must stay within the repo root.");
-  });
-
-  it("rejects preserved gateway artifacts that traverse symlinks", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-gateway-guard-repo-");
-    const outsideRoot = await tempDirs.makeTempDir("qa-gateway-guard-outside-");
-    await mkdir(path.join(repoRoot, ".artifacts"), { recursive: true });
-    await symlink(outsideRoot, path.join(repoRoot, ".artifacts", "qa-e2e"), "dir");
-
-    await expect(
-      testing.assertQaArtifactDirWithinRepo(
-        repoRoot,
-        path.join(repoRoot, ".artifacts", "qa-e2e", "gateway-runtime"),
-      ),
-    ).rejects.toThrow("QA gateway artifact directory must not traverse symlinks.");
-  });
-
-  it("cleans startup temp roots when they are not preserved", async () => {
-    const tempRoot = await tempDirs.makeTempDir("qa-gateway-cleanup-src-");
-    const stagedRoot = await tempDirs.makeTempDir("qa-gateway-cleanup-stage-");
-
-    await writeFile(path.join(tempRoot, "openclaw.json"), "{}", "utf8");
-    await writeFile(path.join(stagedRoot, "marker.txt"), "x", "utf8");
-
-    await testing.cleanupQaGatewayTempRoots({
-      tempRoot,
-      stagedBundledPluginsRoot: stagedRoot,
-    });
-
-    await expectPathMissing(tempRoot);
-    await expectPathMissing(stagedRoot);
-  });
-});
-
-describe("resolveQaControlUiRoot", () => {
-  it("returns the built control ui root when repo assets exist", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-control-ui-root-");
-    const controlUiRoot = path.join(repoRoot, "dist", "control-ui");
-    await mkdir(controlUiRoot, { recursive: true });
-    await writeFile(path.join(controlUiRoot, "index.html"), "<html></html>", "utf8");
-
-    expect(resolveQaControlUiRoot({ repoRoot })).toBe(controlUiRoot);
-  });
-
-  it("returns undefined when control ui is disabled or not built", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-control-ui-root-missing-");
-
-    expect(resolveQaControlUiRoot({ repoRoot })).toBeUndefined();
-    expect(resolveQaControlUiRoot({ repoRoot, controlUiEnabled: false })).toBeUndefined();
-  });
 });
 
 describe("qa bundled plugin dir", () => {
-  it("prefers a built bundled plugin when present", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-bundled-root-");
-    await writeJsonFixture(
-      path.join(repoRoot, "dist", "extensions", "qa-channel", "package.json"),
-      {},
-    );
-    await writeJsonFixture(
-      path.join(repoRoot, "dist-runtime", "extensions", "qa-channel", "package.json"),
-      {},
-    );
-    await writeJsonFixture(path.join(repoRoot, "extensions", "qa-channel", "package.json"), {});
-
-    expect(
-      testing.resolveQaBundledPluginSourceDir({
-        repoRoot,
-        pluginId: "qa-channel",
-      }),
-    ).toBe(path.join(repoRoot, "dist", "extensions", "qa-channel"));
-  });
-
-  it("falls back to the source bundled plugin when no built copy exists", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-bundled-source-root-");
-    await writeJsonFixture(path.join(repoRoot, "extensions", "qa-channel", "package.json"), {});
-
-    expect(
-      testing.resolveQaBundledPluginSourceDir({
-        repoRoot,
-        pluginId: "qa-channel",
-      }),
-    ).toBe(path.join(repoRoot, "extensions", "qa-channel"));
-  });
-
-  it("resolves bundled plugins by manifest id when the directory name differs", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-bundled-manifest-id-root-");
-    await writeJsonFixture(
-      path.join(repoRoot, "dist", "extensions", "kimi-coding", "openclaw.plugin.json"),
-      { id: "kimi", providers: ["kimi"] },
-    );
-    await writeJsonFixture(
-      path.join(repoRoot, "dist", "extensions", "kimi-coding", "package.json"),
-      {},
-    );
-
-    expect(
-      testing.resolveQaBundledPluginSourceDir({
-        repoRoot,
-        pluginId: "kimi",
-      }),
-    ).toBe(path.join(repoRoot, "dist", "extensions", "kimi-coding"));
-  });
-
-  it("uses a source bundled plugin when the built copy is missing CLI metadata", async () => {
-    const repoRoot = await tempDirs.makeTempDir("qa-bundled-cli-metadata-root-");
-    await writeJsonFixture(
-      path.join(repoRoot, "dist", "extensions", "memory-core", "package.json"),
-      {},
-    );
-    await writeJsonFixture(
-      path.join(repoRoot, "dist", "extensions", "memory-core", "openclaw.plugin.json"),
-      { id: "memory-core", kind: "memory" },
-    );
-    await writeJsonFixture(path.join(repoRoot, "extensions", "memory-core", "package.json"), {});
-    await writeJsonFixture(
-      path.join(repoRoot, "extensions", "memory-core", "openclaw.plugin.json"),
-      { id: "memory-core", kind: "memory" },
-    );
-    await writeFile(
-      path.join(repoRoot, "extensions", "memory-core", "cli-metadata.ts"),
-      "export default { id: 'memory-core' };\n",
-      "utf8",
-    );
-
-    expect(
-      testing.resolveQaBundledPluginSourceDir({
-        repoRoot,
-        pluginId: "memory-core",
-      }),
-    ).toBe(path.join(repoRoot, "extensions", "memory-core"));
-  });
-
   it("creates a scoped bundled plugin tree for allowed plugins plus always-allowed runtime facades", async () => {
     const repoRoot = await tempDirs.makeTempDir("qa-bundled-scope-");
     await writeFile(
@@ -2350,7 +1929,7 @@ describe("qa bundled plugin dir", () => {
     await writeFile(path.join(repoRoot, "dist", "shared-chunk-abc123.js"), "export {};\n", "utf8");
     const tempRoot = await tempDirs.makeTempDir("qa-bundled-target-");
 
-    const { bundledPluginsDir, stagedRoot } = await testing.createQaBundledPluginsDir({
+    const { bundledPluginsDir, stagedRoot } = await createQaBundledPluginsDir({
       repoRoot,
       tempRoot,
       allowedPluginIds: ["qa-channel", "memory-core"],
@@ -2439,7 +2018,7 @@ describe("qa bundled plugin dir", () => {
     );
     const tempRoot = await tempDirs.makeTempDir("qa-bundled-mixed-target-");
 
-    const { bundledPluginsDir } = await testing.createQaBundledPluginsDir({
+    const { bundledPluginsDir } = await createQaBundledPluginsDir({
       repoRoot,
       tempRoot,
       allowedPluginIds: ["runtime-only"],
@@ -2486,7 +2065,7 @@ describe("qa bundled plugin dir", () => {
     const tempRoot = await tempDirs.makeTempDir("qa-bundled-invalid-target-");
 
     await expect(
-      testing.createQaBundledPluginsDir({
+      createQaBundledPluginsDir({
         repoRoot,
         tempRoot,
         allowedPluginIds: ["../escape"],
@@ -2503,7 +2082,7 @@ describe("qa bundled plugin dir", () => {
     );
     const tempRoot = await tempDirs.makeTempDir("qa-bundled-external-target-");
 
-    const { bundledPluginsDir } = await testing.createQaBundledPluginsDir({
+    const { bundledPluginsDir } = await createQaBundledPluginsDir({
       repoRoot,
       tempRoot,
       allowedPluginIds: ["external-fixture"],
@@ -2570,7 +2149,7 @@ describe("qa bundled plugin dir", () => {
     await symlink(fakeDepPackageDir, path.join(repoRoot, "node_modules", "fake-dep"), "dir");
     const tempRoot = await tempDirs.makeTempDir("qa-bundled-source-target-");
 
-    const { bundledPluginsDir, stagedRoot } = await testing.createQaBundledPluginsDir({
+    const { bundledPluginsDir, stagedRoot } = await createQaBundledPluginsDir({
       repoRoot,
       tempRoot,
       allowedPluginIds: ["qa-channel"],
@@ -2615,7 +2194,7 @@ describe("qa bundled plugin dir", () => {
     );
 
     await expect(
-      testing.resolveQaOwnerPluginIdsForProviderIds({
+      resolveQaOwnerPluginIdsForProviderIds({
         repoRoot,
         providerIds: ["codex-cli"],
       }),
@@ -2634,7 +2213,7 @@ describe("qa bundled plugin dir", () => {
     );
 
     await expect(
-      testing.resolveQaOwnerPluginIdsForProviderIds({
+      resolveQaOwnerPluginIdsForProviderIds({
         repoRoot,
         providerIds: ["custom-openai"],
         providerConfigs: {
@@ -2688,7 +2267,7 @@ describe("qa bundled plugin dir", () => {
       },
     });
 
-    const overrides = await testing.readQaLiveProviderConfigOverrides({
+    const overrides = await readQaLiveProviderConfigOverrides({
       providerIds: ["custom-openai"],
       env: { OPENCLAW_QA_LIVE_PROVIDER_CONFIG_PATH: configPath },
     });
@@ -2711,7 +2290,7 @@ describe("qa bundled plugin dir", () => {
       },
     });
 
-    const overrides = await testing.readQaLiveProviderConfigOverrides({
+    const overrides = await readQaLiveProviderConfigOverrides({
       providerIds: ["openai"],
       env: { OPENCLAW_QA_LIVE_PROVIDER_CONFIG_PATH: configPath },
     });
@@ -2737,37 +2316,13 @@ describe("qa bundled plugin dir", () => {
       },
     });
 
-    const overrides = await testing.readQaLiveProviderConfigOverrides({
+    const overrides = await readQaLiveProviderConfigOverrides({
       providerIds: ["openai"],
       env: { OPENCLAW_QA_LIVE_PROVIDER_CONFIG_PATH: configPath },
     });
     expect(Object.keys(overrides)).toEqual(["openai"]);
     expect(overrides["openai"]).not.toHaveProperty("baseUrl");
     expect(overrides["openai"]?.api).toBe("openai-responses");
-  });
-
-  it("does not copy OpenAI provider configs for custom OpenAI-compatible runs", async () => {
-    const configPath = await writeTempProviderConfig({
-      models: {
-        providers: {
-          openai: {
-            baseUrl: "https://proxy.example.test/v1",
-            models: [],
-            apiKey: {
-              source: "env",
-              id: "OPENCLAW_LIVE_CODEX_API_KEY",
-            },
-          },
-        },
-      },
-    });
-
-    const overrides = await testing.readQaLiveProviderConfigOverrides({
-      providerIds: ["openai"],
-      env: { OPENCLAW_QA_LIVE_PROVIDER_CONFIG_PATH: configPath },
-    });
-    expect(Object.keys(overrides)).toEqual(["openai"]);
-    expect(overrides.openai?.baseUrl).toBe("https://proxy.example.test/v1");
   });
 
   it("raises the QA runtime host version to the highest allowed plugin floor", async () => {
@@ -2783,7 +2338,7 @@ describe("qa bundled plugin dir", () => {
     });
 
     await expect(
-      testing.resolveQaRuntimeHostVersion({
+      resolveQaRuntimeHostVersion({
         repoRoot,
         allowedPluginIds: ["memory-core", "qa-channel"],
       }),
@@ -2802,7 +2357,7 @@ describe("qa bundled plugin dir", () => {
     });
 
     await expect(
-      testing.resolveQaRuntimeHostVersion({
+      resolveQaRuntimeHostVersion({
         repoRoot,
         allowedPluginIds: ["qa-channel"],
       }),
